requirements.txt
Flask==2.0.1
Flask-Bcrypt==0.7.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0
Werkzeug==2.0.1 